requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
